Moved AnalyticsEvent to core module
