+++ core/build.gradle.kts
@@ -43,6 +43,7 @@ dependencies {
     implementation(Dependencies.Android.material)
     implementation(platform(Dependencies.Firebase.bom))
     implementation(Dependencies.Firebase.analytics)
+    implementation(Dependencies.timber)
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.2")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.3.0")
