+++ online/src/main/java/com/yashovardhan99/healersdiary/online/importFirebase/importFirebaseViewModel.kt
@@ -8,9 +8,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.work.*
 import com.google.firebase.auth.FirebaseUser
+import com.yashovardhan99.core.analytics.AnalyticsEvent
 import com.yashovardhan99.healersdiary.AppDataStore
 import com.yashovardhan99.healersdiary.onboarding.OnboardingViewModel
-import com.yashovardhan99.healersdiary.utils.AnalyticsEvent
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
